Added svg icons and imgs. Added styled components. Added a styled button, not finished.
