--- src/components/Footer.css
@@ -1,6 +1,0 @@
-.footer-img {
-  width: 164px;
-  height: 164px;
-  border-radius: 164px;
-  
-}
