--- src/components/Profile.jsx
@@ -1,18 +1,0 @@
-import "./Profile.css"
-
-export const Profile = () => {
-  return (
-    <section id="main-content">
-      <h3>Hi there I´m</h3> 
-      <h1>Sofia Lennbom</h1>
-      <img className="profile-img"
-      src="assets/sofia_lennbom_black_white.jpg"
-      alt="Profile image of the developer"
-      />
-      <h3>Web Developer with a Background in Quality Assurance, Technical documentation and Training</h3>
-      <p>I’m a developer with a curious mindset, eager to learn and build thoughtful digital experiences. I value quality—not just in code, but in the impact it creates.</p>
-      <p>Loyal and hands-on, I’m a team player who takes initiative and enjoys tackling challenges to deliver meaningful, well-crafted solutions.</p>
-      </section>
-  )
-  
-}
