--- src/components/Tech.jsx
@@ -1,9 +1,0 @@
-export const Tech = () => {
-  return (
-    <div>
-    <h2>Tech</h2>
-    <p>HTML, CSS, Flexbox, CSS Grid, JavaScript, ES6, TypeScript, JSX, React, React Hooks, 
-    Node.js,  Mongo DB, MSSQL, Web Accessibility, APIs, mob-programming, pair-programming,  Git, Jenkins, Jira.</p>
-    </div>
-  )
-}
