+++ src/sections/Footer.css
@@ -1,0 +1,6 @@
+.footer-img {
+  width: 164px;
+  height: 164px;
+  border-radius: 164px;
+  
+}
