+++ src/sections/Nav.css
@@ -1,0 +1,54 @@
+.navigation ul{
+  display: none;
+}  
+
+.hamburger{
+  position: static;
+  /* display: block;   */
+  padding-top: 10px;
+  margin-left: 10px;
+  z-index: 10;
+  }
+
+.navigation ul{
+  /*display: ${hamburgerOpen ? "inline" : "none"};*/
+  background: #191919;
+  color: white;
+  height: 100vh;
+  width: 50vw;        
+  margin-top: 50px;
+  position: absolute;
+  } 
+
+
+/* Desktop Widescreen*/
+@media (min-width: 769px){  
+
+  .navigation{
+    display: block;
+    width: 100%;
+    height: 50px;
+    background-color: #191919;
+    color: white;
+   }
+
+   .navigation ul{
+    display: flex;
+    flex-wrap: wrap;
+    float: right;
+    margin: 20px 0px;
+    padding: 0px 25px;
+  } 
+  
+  .navigation ul li{
+    list-style: none;
+    padding-right: 10px;
+  }
+
+  .hamburger{
+    display: none;
+  }
+
+
+  
+}
