+++ src/sections/Profile.css
@@ -1,0 +1,23 @@
+.profile-img {
+  width: 144px;
+  height: 154px;
+  flex-shrink: 0;
+  border-radius: 12px;
+}
+
+/* Tablet */
+@media (min-width: 501px) and (max-width: 768px){
+  .profile-img {
+    width: 260px;
+    height: 277px;
+  }
+
+}
+
+/* Desktop Widescreen */
+@media (min-width: 769px){
+  .profile-img {
+    width: 418px;
+    height: 418px;
+  }
+}
